+++ sdk/cosmos/azure_data_cosmos/src/cosmos_request.rs
@@ -410,7 +410,7 @@ mod tests {
         let item_options = ItemOptions::default()
             .with_session_token("RequestSession".to_string().into())
             .with_custom_headers(request_custom_headers);
-        let req = CosmosRequest::builder(
+        let mut req = CosmosRequest::builder(
             OperationType::Create,
             ResourceLink::root(ResourceType::Documents),
         )
@@ -418,10 +418,8 @@ mod tests {
         .unwrap();
         item_options.apply_headers(&mut req.headers);
 
-        let mut req_with_client_headers = req;
-        req_with_client_headers
-            .request_context
-            .location_endpoint_to_route = Some("https://example.com/".parse().unwrap());
+        req.request_context.location_endpoint_to_route =
+            Some("https://example.com/".parse().unwrap());
 
         let mut client_custom_headers = std::collections::HashMap::new();
         client_custom_headers.insert(
@@ -431,9 +429,9 @@ mod tests {
 
         let client_options =
             CosmosClientOptions::default().with_custom_headers(client_custom_headers);
-        req_with_client_headers.client_headers(&client_options);
+        client_options.apply_headers(&mut req.headers);
 
-        let raw = req_with_client_headers.into_raw_request();
+        let raw = req.into_raw_request();
         let get_header = |name: &HeaderName| {
             raw.headers()
                 .iter()